refactor: move transform to dataset
